Reuse box-shadow code for drop-shadow filter
